requirements.txt
streamlit==1.45.1
openai-whisper==20240930
